move configuration classes to own project
