+++ Plugins/SpecFlow.Actions.Configuration/ISpecFlowActionJsonLoader.cs
@@ -1,0 +1,7 @@
+﻿namespace SpecFlow.Actions.Configuration
+{
+    public interface ISpecFlowActionJsonLoader
+    {
+        string Load();
+    }
+}
