+++ Plugins/SpecFlow.Actions.Configuration/ISpecFlowActionJsonLocator.cs
@@ -1,0 +1,7 @@
+﻿namespace SpecFlow.Actions.Configuration
+{
+    public interface ISpecFlowActionJsonLocator
+    {
+        string? GetFilePath();
+    }
+}
