+++ Plugins/SpecFlow.Actions.Selenium/SpecFlow.Actions.Selenium/RuntimePlugin.cs
@@ -16,8 +16,6 @@ namespace SpecFlow.Actions.Selenium
 
         private void RuntimePluginEvents_CustomizeScenarioDependencies(object sender, CustomizeScenarioDependenciesEventArgs e)
         {
-            e.ObjectContainer.RegisterTypeAs<SpecFlowActionJsonLocator, ISpecFlowActionJsonLocator>();
-            e.ObjectContainer.RegisterTypeAs<SpecFlowActionJsonLoader, ISpecFlowActionJsonLoader>();
             e.ObjectContainer.RegisterTypeAs<SeleniumConfiguration, ISeleniumConfiguration>();
         }
     }
