+++ Plugins/SpecFlow.Actions.Selenium/SpecFlow.Actions.Selenium/SeleniumConfiguration.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using SpecFlow.Actions.Configuration;
+using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
