+++ Plugins/SpecFlow.Actions.Selenium/SpecFlow.Actions.Selenium/SpecFlow.Actions.Selenium.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="System.Text.Json" Version="5.0.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\SpecFlow.Actions.Configuration\SpecFlow.Actions.Configuration.csproj" />
+  </ItemGroup>
+
 </Project>
